Revert "Refactor(bumsoo/letter):  쪽지 보내기 기능 1번 2번 리팩토링"
